backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# OpenAI
openai>=1.12.0

# Database
aiosqlite>=0.19.0

# Validation
pydantic>=2.5.3

# Utilities
httpx>=0.26.0
